Move PersonalDetails to its own file.
