+++ app/services/testdata/GeneratorConfig.scala
@@ -18,7 +18,7 @@ package services.testdata
 
 import model.ApplicationStatuses
 import model.EvaluationResults.Result
-import model.PersistedObjects.PersonalDetails
+import model.persisted.PersonalDetails
 import org.joda.time.LocalDate
 import org.joda.time.format.DateTimeFormat
 import services.testdata.faker.DataFaker.Random
